ferrompi

Module slurm

SLURM scheduler environment helpers.

These functions read SLURM environment variables to determine job topology. They return None if the variable is not set (e.g., when not running under SLURM).

Environment Variables

Function	Variable	Description

job_id()	SLURM_JOB_ID	Unique job identifier

local_rank()	SLURM_LOCALID	Task ID relative to this node

local_size()	SLURM_NTASKS_PER_NODE	Number of tasks on this node

num_nodes()	SLURM_NNODES	Total number of nodes

cpus_per_task()	SLURM_CPUS_PER_TASK	CPUs allocated per task

node_name()	SLURM_NODENAME	Name of this compute node

node_list()	SLURM_NODELIST	Compact list of allocated nodes

Functions

cpus_per_task

Get the number of CPUs per task.

is_slurm_job

Check if running under SLURM job scheduler.

job_id

Get the SLURM job ID.

local_rank

Get the local (intra-node) rank of this process.

local_size

Get the number of tasks per node.

node_list

Get the compact node list string (e.g., “node[001-004]”).

node_name

Get the name of the compute node this process is running on.

num_nodes

Get the total number of nodes allocated.